requirements.txt
numpy==2.0.2
scikit-learn
joblib
mediapipe
opencv-python
Pillow
tk
gradio
watchdog
azure-cognitiveservices-speech